TW-1381: move send_file_dialog component into a folder
